+++ doc/source/doxygen-docs/changelog.md
@@ -24,6 +24,8 @@
     - Optimization: mrpt::maps::CPointsMap::insertAnotherMap() avoids matrix multiplication if SE(3) identity is passed as insertion pose.
   - \ref mrpt_nav_grp
     - mrpt::nav::CParameterizedTrajectoryGenerator::initTPObstacleSingle() now always initializes to the maximum free distance, instead of saturating free space when heading to a target waypoint.
+  - \ref mrpt_obs_grp
+    - Fix const-correctness of mrpt::obs::CObservation::unload() for consistency with load().
   - \ref mrpt_poses_grp
     - New function mrpt::poses::sensor_poses_from_yaml()
   - \ref mrpt_serialization_grp